+++ cli/build.gradle.kts
@@ -8,9 +8,13 @@ repositories {
 }
 
 dependencies {
+    implementation(platform("org.jetbrains.kotlin:kotlin-bom"))
+    implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
+
     implementation(projects.pbCoroutinesCore)
-    implementation(projects.pbCoroutinesKafka)
     implementation(projects.pbCoroutinesRetry)
+    implementation(projects.pbCoroutinesKafka)
+    implementation(projects.pbCoroutinesKafkaRetry)
 
     implementation(libs.bundles.logback)
     implementation(libs.bundles.kotlinxCoroutines)
