+++ pb-coroutines-kafka-retry/build.gradle.kts
@@ -1,13 +1,7 @@
-plugins {
-    kotlin("jvm")
-    `java-library`
-}
-
-repositories {
-    mavenCentral()
-}
-
 dependencies {
+    implementation(libs.kafka)
+
     implementation(projects.pbCoroutinesCore)
     implementation(projects.pbCoroutinesKafka)
+    implementation(projects.pbCoroutinesRetry)
 }
