+++ pb-coroutines-kafka-retry/src/main/kotlin/io/provenance/kafka/coroutines/retry/flow/FlowExtensions.kt
@@ -1,5 +1,6 @@
 package io.provenance.kafka.coroutines.retry.flow
 
+import io.provenance.coroutines.retry.flow.FlowProcessor
 import io.provenance.kafka.UnAckedConsumerRecord
 import io.provenance.kafka.coroutines.retry.tryOnEach
 import org.apache.kafka.clients.consumer.ConsumerRecord
