+++ pb-coroutines-kafka-retry/src/main/kotlin/io/provenance/kafka/coroutines/retry/flow/KafkaFlowRetry.kt
@@ -1,15 +1,18 @@
 package io.provenance.kafka.coroutines.retry.flow
 
+import io.provenance.coroutines.retry.store.RetryRecord
+import io.provenance.coroutines.retry.flow.FlowRetry
+import io.provenance.coroutines.retry.store.RetryRecordStore
 import io.provenance.kafka.coroutines.retry.DEFAULT_RECORD_REPROCESS_GROUP_SIZE
 import io.provenance.kafka.coroutines.retry.KAFKA_RETRY_ATTEMPTS_HEADER
-import io.provenance.kafka.coroutines.retry.store.RetryRecord
-import io.provenance.kafka.coroutines.retry.store.RetryRecordStore
-import io.provenance.kafka.coroutines.retry.store.setHeader
 import io.provenance.kafka.coroutines.retry.toByteArray
 import java.time.OffsetDateTime
 import kotlinx.coroutines.flow.asFlow
 import mu.KotlinLogging
 import org.apache.kafka.clients.consumer.ConsumerRecord
+import org.apache.kafka.common.header.Header
+import org.apache.kafka.common.header.Headers
+import org.apache.kafka.common.header.internals.RecordHeader
 
 /**
  * Retry a flow of kafka records.
@@ -60,5 +63,20 @@ open class KafkaFlowRetry<K, V>(
 
         log.debug { "processing key:${item.key()} attempt:$attempt source:${item.topic()}-${item.partition()}" }
         handler(item.setHeader(KAFKA_RETRY_ATTEMPTS_HEADER, attempt.toByteArray()))
+    }
+
+    private fun <K, V> ConsumerRecord<K, V>.setHeader(key: String, value: ByteArray): ConsumerRecord<K, V> = apply {
+        fun Headers.addOrUpdate(header: Header): Headers {
+            val h = find { it.key() == header.key() }
+            if (h == null) {
+                add(header)
+            } else {
+                remove(header.key())
+                add(header.key(), header.value())
+            }
+            return this
+        }
+
+        headers().addOrUpdate(RecordHeader(key, value))
     }
 }
