+++ pb-coroutines-retry/src/main/kotlin/io/provenance/coroutines/retry/flow/FlowProcessor.kt
@@ -1,4 +1,4 @@
-package io.provenance.kafka.coroutines.retry.flow
+package io.provenance.coroutines.retry.flow
 
 /**
  * Callback hook for submitting an item into a retry flow.
