+++ pb-coroutines-retry/src/main/kotlin/io/provenance/coroutines/retry/flow/RetryFlow.kt
@@ -1,20 +1,16 @@
-package io.provenance.kafka.coroutines.retry.flow
+package io.provenance.coroutines.retry.flow
 
+import io.provenance.coroutines.retry.RetryStrategy
+import io.provenance.coroutines.retry.defaultRetryStrategies
+import io.provenance.coroutines.retry.invert
+import io.provenance.coroutines.retry.store.RetryRecord
 import io.provenance.coroutines.tryMap
-import io.provenance.kafka.coroutines.retry.RetryStrategy
-import io.provenance.kafka.coroutines.retry.defaultRetryStrategies
-import io.provenance.kafka.coroutines.retry.invert
-import io.provenance.kafka.coroutines.retry.store.RetryRecord
 import java.time.OffsetDateTime
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
 import kotlin.time.ExperimentalTime
 import kotlin.time.toJavaDuration
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.channels.ProducerScope
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.channelFlow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onStart
@@ -62,19 +58,5 @@ fun <T> retryFlow(
                 }
                 .collect()
         }
-    }
-}
-
-/**
- * Create a polling [Flow].
- *
- * @param pollInterval The interval to wait between calls to [block].
- * @param block Lambda returning the next element to be emitted, or null if nothing is ready.
- */
-@OptIn(ExperimentalCoroutinesApi::class)
-inline fun <T> pollingFlow(pollInterval: Duration, crossinline block: suspend ProducerScope<T>.() -> Unit): Flow<T> = channelFlow {
-    while (!isClosedForSend) {
-        block(this)
-        delay(pollInterval.inWholeMilliseconds)
     }
 }
