--- pb-coroutines-retry/src/main/kotlin/io/provenance/kafka/coroutines/retry/store/ConsumerRecordRetry.kt
@@ -1,36 +1,0 @@
-package io.provenance.kafka.coroutines.retry.store
-
-import java.time.OffsetDateTime
-import org.apache.kafka.clients.consumer.ConsumerRecord
-import org.apache.kafka.common.header.Header
-import org.apache.kafka.common.header.Headers
-import org.apache.kafka.common.header.internals.RecordHeader
-
-open class RetryRecord<T>(
-    val data: T,
-    val attempt: Int,
-    val lastAttempted: OffsetDateTime,
-) {
-    fun copy(
-        data: T = this.data,
-        attempt: Int = this.attempt,
-        lastAttempted: OffsetDateTime = this.lastAttempted
-    ) = RetryRecord(data, attempt, lastAttempted)
-
-    override fun toString(): String = "RetryRecord(attempt:$attempt lastAttempted:$lastAttempted data:$data)"
-}
-
-private fun Headers.addOrUpdate(header: Header): Headers {
-    val h = find { it.key() == header.key() }
-    if (h == null) {
-        add(header)
-    } else {
-        remove(header.key())
-        add(header.key(), header.value())
-    }
-    return this
-}
-
-fun <K, V> ConsumerRecord<K, V>.setHeader(key: String, value: ByteArray): ConsumerRecord<K, V> = apply {
-    headers().addOrUpdate(RecordHeader(key, value))
-}
